refactor, rename: 프로젝트 상세 페이지 SSR 방식으로 변경 및 CSR 컴포넌트 client 폴더로 이동
