+++ src/components/projectDetail/ProjectDetail.tsx
@@ -1,15 +1,7 @@
-"use client"
-
-import { useState, useRef, useEffect } from "react"
-import { Share2, ChevronLeft, ChevronRight, X, Copy, Check } from "lucide-react"
 import ProjectTabs from "./ProjectTabs"
-
-// 상단에 useApi 훅 import 추가
-import { useApi } from "@/hooks/useApi"
-import { useParams } from "next/navigation"
-
-import { GhostButton, OutlineButton, PrimaryButton } from "../common/Button"
-import gsap from "gsap"
+import ImageCarousel from "./client/ImageCarousel"
+import SharePopover from "./client/SharePopover"
+import DonateButton from "./client/DonateButton"
 
 // ProjectData 인터페이스 추가
 interface ProjectData {
@@ -39,91 +31,24 @@ interface ProjectData {
   }[]
 }
 
-const ProjectDetail = () => {
-  const [currentImageIndex, setCurrentImageIndex] = useState(0)
-  const [prevImageIndex, setPrevImageIndex] = useState<number | null>(null)
-  const [isSliding, setIsSliding] = useState(false)
-  const [slideDirection, setSlideDirection] = useState<'left' | 'right'>("right")
-  const imageContainerRef = useRef<HTMLDivElement>(null)
-  const metadataContainerRef = useRef<HTMLDivElement>(null)
+interface ProjectDetailProps {
+  projectData: ProjectData | null
+  isLoading: boolean
+  error: string | null
+}
 
-  // API 관련 상태 추가
-  const { apiCall, isLoading: apiLoading } = useApi()
-  const [projectData, setProjectData] = useState<ProjectData | null>(null)
-  const [isLoading, setIsLoading] = useState(true)
-  const [error, setError] = useState<string | null>(null)
-
-  // URL에서 프로젝트 ID 가져오기
-  const { id: projectId } = useParams()
-
-  // 후원 가능 여부 확인 함수
+const ProjectDetail = ({ projectData, isLoading, error }: ProjectDetailProps) => {
+  // SSR에서 사용 가능한 순수 함수만 남김
   const canDonate = (projectData: ProjectData | null): boolean => {
     if (!projectData?.productResponses || projectData.productResponses.length === 0) {
       return false
     }
-
-    // 모든 상품의 재고가 0인지 확인
     const hasAvailableStock = projectData.productResponses.some((product) => product.stock > 0)
     return hasAvailableStock
   }
 
-  const [showSharePopover, setShowSharePopover] = useState(false)
-  const [copySuccess, setCopySuccess] = useState(false)
-
-  // 이미지 캐러셀에 사용할 이미지 DOM 배열을 위한 ref
-  const imageRefs = useRef<(HTMLImageElement | null)[]>([])
-
-  // 프로젝트 데이터 가져오기
-  useEffect(() => {
-    const fetchProjectData = async () => {
-      try {
-        setIsLoading(true)
-        setError(null)
-
-        const { data, error } = await apiCall<ProjectData>(`/api/project/${projectId}`, "GET")
-
-        if (error) {
-          throw new Error(error)
-        }
-
-        if (data) {
-          console.log(data)
-          setProjectData(data)
-        }
-      } catch (err) {
-        setError(err instanceof Error ? err.message : "프로젝트 정보를 불러오는 중 오류가 발생했습니다.")
-        console.error("프로젝트 정보 불러오기 오류:", err)
-      } finally {
-        setIsLoading(false)
-      }
-    }
-
-    if (projectId) {
-      fetchProjectData()
-    }
-  }, [projectId, apiCall])
-
-  // 사용할 이미지 배열 결정
   const images = projectData?.imageUrls?.length ? projectData.imageUrls : []
 
-  const slideTo = (dir: 'left' | 'right') => {
-    if (isSliding || images.length <= 1) return
-    setIsSliding(true)
-    setSlideDirection(dir)
-    setPrevImageIndex(currentImageIndex)
-    let nextIdx
-    if (dir === 'right') {
-      nextIdx = (currentImageIndex + 1) % images.length
-    } else {
-      nextIdx = (currentImageIndex - 1 + images.length) % images.length
-    }
-    setCurrentImageIndex(nextIdx)
-  }
-
-  const nextImage = () => slideTo('right')
-  const prevImage = () => slideTo('left')
-
-  // 날짜 포맷 함수
   const formatDate = (dateString: string) => {
     if (!dateString) return ""
     const date = new Date(dateString)
@@ -136,7 +61,6 @@ const ProjectDetail = () => {
       .replace(/\. /g, ". ")
   }
 
-  // 남은 일수 계산 함수
   const calculateDaysLeft = (endDate: string) => {
     if (!endDate) return 0
     const end = new Date(endDate)
@@ -146,86 +70,10 @@ const ProjectDetail = () => {
     return diffDays
   }
 
-  // 달성률 계산 함수
   const calculateAchievementRate = (goal: number, current: number) => {
     if (!goal || goal === 0) return 0
     return Math.floor((current / goal) * 100)
   }
-
-  // URL 복사 함수
-  const copyToClipboard = async () => {
-    try {
-      const currentUrl = window.location.href
-      await navigator.clipboard.writeText(currentUrl)
-      setCopySuccess(true)
-      setTimeout(() => {
-        setCopySuccess(false)
-        setShowSharePopover(false)
-      }, 3000)
-    } catch (err) {
-      console.error("URL 복사 실패:", err)
-      // fallback for older browsers
-      const textArea = document.createElement("textarea")
-      textArea.value = window.location.href
-      document.body.appendChild(textArea)
-      textArea.select()
-      document.execCommand("copy")
-      document.body.removeChild(textArea)
-      setCopySuccess(true)
-      setTimeout(() => {
-        setCopySuccess(false)
-        setShowSharePopover(false)
-      }, 2000)
-    }
-  }
-
-  // 팝오버 외부 클릭 시 닫기
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      const target = event.target as Element
-      if (showSharePopover && !target.closest(".relative")) {
-        setShowSharePopover(false)
-      }
-    }
-
-    if (showSharePopover) {
-      document.addEventListener("mousedown", handleClickOutside)
-    }
-
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside)
-    }
-  }, [showSharePopover])
-
-  // 슬라이드 애니메이션 useEffect
-  useEffect(() => {
-    if (prevImageIndex === null || prevImageIndex === currentImageIndex) return
-    const prevImg = imageRefs.current[0]
-    const currImg = imageRefs.current[1]
-    if (!prevImg || !currImg) {
-      setIsSliding(false)
-      setPrevImageIndex(null)
-      return
-    }
-    // 초기 위치 세팅
-    gsap.set(prevImg, { xPercent: 0, zIndex: 1 })
-    gsap.set(currImg, { xPercent: slideDirection === 'right' ? 100 : -100, zIndex: 2 })
-    // 애니메이션
-    gsap.to(prevImg, {
-      xPercent: slideDirection === 'right' ? -100 : 100,
-      duration: 0.5,
-      ease: 'power2.inOut',
-      onComplete: () => {
-        setIsSliding(false)
-        setPrevImageIndex(null)
-      }
-    })
-    gsap.to(currImg, {
-      xPercent: 0,
-      duration: 0.5,
-      ease: 'power2.inOut',
-    })
-  }, [currentImageIndex, prevImageIndex, slideDirection])
 
   return (
     <div className="mx-auto max-w-6xl pb-20">
@@ -263,7 +111,7 @@ const ProjectDetail = () => {
       {error && (
         <div className="rounded-xl bg-red-50 p-4 text-red-500 my-4">
           <p>{error}</p>
-          <button onClick={() => window.location.reload()} className="mt-2 text-sm underline">
+          <button onClick={() => typeof window !== "undefined" && window.location.reload()} className="mt-2 text-sm underline">
             다시 시도
           </button>
         </div>
@@ -274,92 +122,12 @@ const ProjectDetail = () => {
         <div className="grid grid-cols-1 gap-8 md:grid-cols-2">
           {/* 왼쪽: 이미지 영역 */}
           <div className="flex flex-col">
-            {/* 이미지 표시 영역 (캐러셀) */}
-            <div ref={imageContainerRef} className="relative aspect-square w-full overflow-hidden rounded-3xl">
-              {images.length > 0 ? (
-                <>
-                  {/* 두 장의 이미지만 렌더링: prev, current */}
-                  {prevImageIndex !== null && prevImageIndex !== currentImageIndex && (
-                    <img
-                      data-cy="project-image"
-                      key={prevImageIndex}
-                      ref={el => { imageRefs.current[0] = el }}
-                      src={images[prevImageIndex] || "/placeholder.svg"}
-                      alt={projectData?.title || `프로젝트 이미지 ${prevImageIndex + 1}`}
-                      className="absolute top-0 left-0 h-full w-full object-cover"
-                      style={{ zIndex: 1 }}
-                    />
-                  )}
-                  <img
-                    data-cy="project-image"
-                    key={currentImageIndex}
-                    ref={el => { imageRefs.current[1] = el }}
-                    src={images[currentImageIndex] || "/placeholder.svg"}
-                    alt={projectData?.title || `프로젝트 이미지 ${currentImageIndex + 1}`}
-                    className="absolute top-0 left-0 h-full w-full object-cover"
-                    style={{ zIndex: 2 }}
-                  />
-                  {/* 좌우 버튼 */}
-                  {images.length > 1 && (
-                    <>
-                      <button
-                        onClick={prevImage}
-                        className="absolute left-4 top-1/2 -translate-y-1/2 rounded-full bg-white/70 p-2 text-sub-gray shadow-md hover:bg-white z-4"
-                        aria-label="이전 이미지"
-                        disabled={isSliding}
-                      >
-                        <ChevronLeft className="h-6 w-6" />
-                      </button>
-                      <button
-                        onClick={nextImage}
-                        className="absolute right-4 top-1/2 -translate-y-1/2 rounded-full bg-white/70 p-2 text-sub-gray shadow-md hover:bg-white z-4"
-                        aria-label="다음 이미지"
-                        disabled={isSliding}
-                      >
-                        <ChevronRight className="h-6 w-6" />
-                      </button>
-                    </>
-                  )}
-                </>
-              ) : (
-                <>
-                  <div className="h-full w-full bg-gray-100 flex items-center justify-center">
-                    <div className="text-center">
-                      <p className="text-gray-500 text-xl font-medium">이미지 없음</p>
-                    </div>
-                  </div>
-                </>
-              )}
-            </div>
-
-            {/* 이미지 목록 */}
-            {images.length > 0 && (
-              <div className="mt-4">
-                <div className="flex justify-center gap-4">
-                  {images.map((img, idx) => (
-                    <div
-                      key={idx}
-                      className={`cursor-pointer overflow-hidden rounded-xl transition-all duration-200 ${
-                        currentImageIndex === idx ? "ring-4 ring-main-color" : "hover:ring-2 hover:ring-main-color"
-                      }`}
-                      onClick={() => !isSliding && setCurrentImageIndex(idx)}
-                    >
-                      <div className="relative h-20 w-20">
-                        <img
-                          src={img || "/placeholder.svg"}
-                          alt={`프로젝트 이미지 ${idx + 1}`}
-                          className="w-full h-full object-cover"
-                        />
-                      </div>
-                    </div>
-                  ))}
-                </div>
-              </div>
-            )}
+            {/* 이미지 캐러셀 분리 */}
+            <ImageCarousel images={images} title={projectData?.title || "프로젝트 이미지"} />
           </div>
 
           {/* 오른쪽: 메타데이터 영역 */}
-          <div ref={metadataContainerRef} className="flex flex-col justify-center px-4 md:px-0 text-center md:text-left">
+          <div className="flex flex-col justify-center px-4 md:px-0 text-center md:text-left">
             <div>
               <h1 className="text-lg md:text-3xl font-bold mb-2 text-center md:text-left">{projectData?.title || "프로젝트 제목"}</h1>
               <h2 className="text-sm md:text-2xl text-sub-gray mb-6 break-words leading-tight text-center md:text-left">{projectData?.description || "프로젝트 설명"}</h2>
@@ -370,19 +138,15 @@ const ProjectDetail = () => {
                   <span className="text-2xl md:text-5xl font-bold text-main-color">
                     {projectData?.totalAmount?.toLocaleString() || "0"}원
                   </span>
-                  {/* 데스크톱: 기존처럼 목표금액 옆에 표기 */}
                   <span className="ml-2 text-base md:text-xl text-sub-gray hidden md:inline">
                     / {projectData?.goalAmount?.toLocaleString() || "0"}원
                   </span>
                 </div>
               </div>
 
-              {/* 달성률 표기 */}
-              {/* 데스크톱: 기존 한 줄 표기 */}
               <p className="text-base md:text-lg font-medium text-main-color hidden md:block mb-4 md:mb-6 text-left">
                 목표 금액의 {calculateAchievementRate(projectData?.goalAmount || 0, projectData?.totalAmount || 0)}% 달성
               </p>
-              {/* 모바일: 두 줄 표기 */}
               <div className="block md:hidden text-main-color text-base font-medium mb-4 text-center">
                 <div>
                   목표금액({projectData?.goalAmount?.toLocaleString() || "0"}원)의
@@ -416,77 +180,8 @@ const ProjectDetail = () => {
 
             {/* 공유하기, 후원하기 버튼 영역 */}
             <div className="mt-6 md:mt-8 flex flex-col md:flex-row items-stretch md:items-center gap-2 md:justify-end text-center md:text-right">
-              {/* 공유하기 버튼을 relative 컨테이너로 감싸기 */}
-              <div className="relative w-full md:w-1/3">
-                <OutlineButton
-                  className="w-full flex items-center justify-center gap-2 rounded-xl border border-gray-300 bg-white py-3 md:py-4 text-base md:text-xl text-gray-700 hover:bg-gray-50"
-                  onClick={() => setShowSharePopover(!showSharePopover)}
-                >
-                  <Share2 className="h-5 w-5 md:h-6 md:w-6" />
-                  공유하기
-                </OutlineButton>
-
-                {/* 공유하기 팝오버 - 트랜지션 애니메이션 적용, 항상 렌더링 */}
-                <div
-                  className={`absolute top-full left-0 mt-2 w-64 bg-white rounded-lg shadow-lg border border-gray-200 p-4 z-10 transition-all duration-75 ease-out
-                    ${showSharePopover ? 'opacity-100 scale-100 pointer-events-auto' : 'opacity-0 scale-95 pointer-events-none'}`}
-                  style={{transformOrigin: 'top left'}}
-                >
-                  <div className="flex items-center justify-between mb-3">
-                    <h3 className="font-medium text-gray-900">프로젝트 공유하기</h3>
-                    <GhostButton onClick={() => setShowSharePopover(false)} className="p-1 rounded-full">
-                      <X className="h-4 w-4" />
-                    </GhostButton>
-                  </div>
-
-                  <div className="space-y-3">
-                    <div className="flex items-center gap-2 p-2 bg-gray-50 rounded border text-sm text-gray-600">
-                      <span className="flex-1 truncate">
-                        {typeof window !== "undefined" ? window.location.href : ""}
-                      </span>
-                    </div>
-
-                    <PrimaryButton
-                      onClick={copyToClipboard}
-                      className="w-full flex items-center justify-center gap-2 py-2 px-4 rounded-lg transition-colors"
-                      disabled={copySuccess}
-                    >
-                      {copySuccess ? (
-                        <>
-                          <Check className="h-4 w-4" />
-                          복사 완료!
-                        </>
-                      ) : (
-                        <>
-                          <Copy className="h-4 w-4" />
-                          URL 복사하기
-                        </>
-                      )}
-                    </PrimaryButton>
-                  </div>
-                </div>
-              </div>
-
-              {/* 후원하기 버튼 - 상품이 없거나 모든 재고가 0이면 비활성화 */}
-              {canDonate(projectData) ? (
-                <PrimaryButton
-                  className="w-full md:w-2/3 rounded-xl bg-main-color px-4 md:px-8 py-3 md:py-4 text-base md:text-xl font-bold text-white hover:bg-secondary-color-dark"
-                  onClick={() => {
-                    const event = new CustomEvent("toggleDonateDock")
-                    window.dispatchEvent(event)
-                  }}
-                >
-                  후원하기
-                </PrimaryButton>
-              ) : (
-                <PrimaryButton
-                  className="w-full md:w-2/3 rounded-xl px-4 md:px-8 py-3 md:py-4 text-base md:text-xl font-bold"
-                  disabled
-                  data-cy="donate-disabled"
-                >
-                  후원 불가
-                </PrimaryButton>
-              )}
+              <SharePopover />
+              <DonateButton canDonate={canDonate(projectData)} />
             </div>
           </div>
         </div>
